+++ app/controllers/shifts_controller.rb
@@ -5,21 +5,20 @@ class ShiftsController < ApplicationController
   before_action :set_member
   before_action :set_params, only: [:create, :update]
   def index
-      @shifts = Shift.where(store_id: params[:store_id])
-      @message = "シフト提出状況"
-      # @members = Member.where(store_id: @store.id)
+    @shifts = Shift.where(store_id: params[:store_id])
+    @message = 'シフト提出状況'
+    # @members = Member.where(store_id: @store.id)
 
-      @members = @store.members.includes(:store)
-      shifts = Shift.where(store_id: params[:store_id])
-      @shifts = shifts.where('workday like ?',"#{@next.year}-#{@next.month}%")
-      tables = Table.where(store_id: (params[:store_id]))
-      @tables = tables.where('workday like ?',"#{@next.year}-#{@next.month}%")  
-      
+    @members = @store.members.includes(:store)
+    shifts = Shift.where(store_id: params[:store_id])
+    @shifts = shifts.where('workday like ?', "#{@next.year}-#{@next.month}%")
+    tables = Table.where(store_id: params[:store_id])
+    @tables = tables.where('workday like ?', "#{@next.year}-#{@next.month}%")
   end
 
   def new
     @shifts = @store.shifts.includes(:user)
-    @message = "希望シフトを提出します"
+    @message = '希望シフトを提出します'
     @shift_where = Shift.where(user_id: current_user.id, store_id: @store.id)
     @shift = Shift.new
   end
@@ -36,7 +35,7 @@ class ShiftsController < ApplicationController
   end
 
   def edit
-    @message = "シフトを変更します"
+    @message = 'シフトを変更します'
     @shift = Shift.find(params[:id])
   end
 
@@ -48,14 +47,11 @@ class ShiftsController < ApplicationController
 
   def destroy
     shift = Shift.find(params[:id])
-    if shift.user_id == current_user.id
-      shift.destroy
-    end
+    shift.destroy if shift.user_id == current_user.id
     redirect_to action: :new
   end
 
   def show
-    
   end
 
   private
@@ -86,22 +82,20 @@ class ShiftsController < ApplicationController
   def set_date
     @d = Date.today
     @year = @d.year
-    @month = @d.next_month.month #シフト提出は翌月分になるため、monthは来月の月を定義する
+    @month = @d.next_month.month # シフト提出は翌月分になるため、monthは来月の月を定義する
     # @monthが１月の場合、@yearが１年繰り上がるので、来年を表示できるように設定
-    if @month == 1
-      @year += 1
-    end
+    @year += 1 if @month == 1
     # ビューファイル用
-    @all_days = Date.new(@year,@month,-1) #該当の月の存在する日数分の値を取り出す
-    @count = 1 #ビューファイルで日付としてしようするため、定義する
+    @all_days = Date.new(@year, @month, -1) # 該当の月の存在する日数分の値を取り出す
+    @count = 1 # ビューファイルで日付としてしようするため、定義する
     @store = Store.find(params[:store_id])
-    @ja = %w(日 月 火 水 木 金 土)
+    @ja = %w[日 月 火 水 木 金 土]
   end
 
   def set_month
     @today = Date.today
     @next = @today.next_month
-    @day = Date.new(@next.year,@next.month,-1)
+    @day = Date.new(@next.year, @next.month, -1)
     @day_count = 0
     # @test = Date.new(@next.year,@next.month,@day_count)
   end
@@ -114,7 +108,4 @@ class ShiftsController < ApplicationController
     @user_id = current_user.id
     @store_id = params[:store_id]
   end
- 
-
- 
 end
